refactor: simplify test session factory
